pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-cls"
version = "0.1.0"
description = "Document Classification & Metadata Extraction System"
requires-python = ">=3.12"
dependencies = [
    # Core schemas & config
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # OCR — Azure Document Intelligence
    "azure-ai-documentintelligence>=1.0",
    "azure-identity>=1.17",

    # Local image loading & quality assessment (run before Azure DI)
    "pdf2image>=1.17",
    "Pillow>=10.0",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",

    # LLM orchestration (includes AzureChatOpenAI)
    "langchain-core>=0.3",
    "langchain-openai>=0.3",

    # Structured logging
    "structlog>=24.1",

    # Retry with backoff
    "tenacity>=8.2",

    # Data export (built-in json/csv modules used)

    # CLI
    "click>=8.1",

    # Environment variable loading
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "schemas*", "ocr*", "classifier*", "confidence*", "pipeline*", "exporters*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
